codacy: move listener to own file to reduce file complexity
